test(md-to-office): split test_md_to_office.py into topic modules (closes #330)
The file was 1262 LOC with 9 test classes spanning four distinct test
layers (per the original docstring: resolve chain, orchestrator,
brand scanner, end-to-end render). Split for readability and to bring
every file under the 500 LOC oversized threshold flagged by tech-lead's
audit.

New layout (each file < 500 LOC):

- _md_to_office_helpers.py            — shared script paths + run()
- test_md_to_office_resolve.py        — 5-priority resolve-template chain
- test_md_to_office_orchestrator.py   — md-to-office.sh target gating
- test_md_to_office_scan.py           — scan-brand.py signal extraction
- test_md_to_office_init.py           — generate-templates.py + init-brand.sh
- test_md_to_office_render.py         — render-docx/pptx + markdown parser
- test_md_to_office_styling.py        — docx professional styling presets

test_md_to_office.py becomes a thin aggregator that imports every test
class so unittest.main() discovers them. CI's
`python3 claude-skills/tests/test_md_to_office.py` invocation keeps
working unchanged. Each topic module is also independently runnable.

Test count is unchanged (97 → 97 across the seven modules).
